Move Friendica\Core\NotificationsManager to Friendica\Model\Notify
